Refactoring
-Changed 'Game' to 'Blackjack'
-Changed 'Main' to 'Game'
-Removed playmode
